requirements.txt
flask==3.1.0
anthropic>=0.50.0
groq>=0.9.0
python-dotenv==1.0.1
requests>=2.31.0
